Refactoring - remove same code from finders
